connected with finance and library services

+++ student-portal/src/main/java/com/shramaner/studentPortal/model/Student.java
@@ -27,7 +27,7 @@ public class Student {
     @Column(nullable = false)
     private Integer age;
     @Column(nullable = false, unique = true)
-    private Integer phone;
+    private long phone;
